requirements.txt
matplotlib==3.5.2
numpy==1.23.0
pandas==1.4.3
pip==22.2.1
python-dotenv==0.20.0
seaborn==0.11.2
tensorflow==2.9.1
